+++ simpleimages/callers.py
@@ -1,3 +1,8 @@
+from __future__ import absolute_import
+
+import dill
+
+
 def default(function, *args, **kwargs):
     '''
     Calls ``function`` with any passed in ``args`` and ``kwargs.
@@ -14,9 +19,10 @@ def _no_action(function, *args, **kwargs):
 
 def celery(function, *args, **kwargs):
     '''
-    Calls ``function`` asynchronously by creating a
-    `shared-task <http://docs.celeryproject.org/en/latest/django/first-steps-with-django.html#using-the-shared-task-decorator>`_
-    with celery.
+    Calls ``function`` asynchronously by creating a pickling it and
+    calling it in a task.
     '''
-    from celery import shared_task
-    shared_task(function).delay(*args, **kwargs)
+    from .tasks import dill_callable
+
+    dilled_function = dill.dumps(function)
+    dill_callable.delay(dilled_function, *args, **kwargs)